+++ .gitignore
@@ -13,7 +13,7 @@
 *.out
 coverage.txt
 
-/goclitem
+/test/testapp
 
 libexec/
 bin/
