+++ .goreleaser.yml
@@ -1,7 +1,7 @@
 # This is an example goreleaser.yaml file with some sane defaults.
 # Make sure to check the documentation at http://goreleaser.com
 builds:
-- main: ./cmd/goclitem/main.go
+- main: ./cmd/gobail/main.go
   env:
   - CGO_ENABLED=0
   - GOPROXY=https://proxy.golang.org
