+++ go.mod
@@ -1,14 +1,10 @@
-module github.com/dnnrly/goclitem
+module github.com/dnnrly/gobail
 
 go 1.13
 
 require (
-	github.com/armon/consul-api v0.0.0-20180202201655-eb2c6b5be1b6 // indirect
-	github.com/coreos/go-etcd v2.0.0+incompatible // indirect
-	github.com/cpuguy83/go-md2man v1.0.10 // indirect
-	github.com/cucumber/godog v0.11.0
-	github.com/spf13/cobra v1.4.0
-	github.com/stretchr/testify v1.6.1
-	github.com/ugorji/go/codec v0.0.0-20181204163529-d75b2dcb6bc8 // indirect
-	github.com/xordataexchange/crypt v0.0.3-0.20170626215501-b2862e3d0a77 // indirect
+	github.com/cucumber/godog v0.15.0
+	github.com/kr/text v0.2.0 // indirect
+	github.com/spf13/cobra v1.7.0
+	github.com/stretchr/testify v1.8.2
 )
