--- test/basic.bats
@@ -1,8 +1,0 @@
-BIN=./goclitem
-
-@test "Can run the application" {
-    run ${BIN}
-    echo $output
-    [ $status -eq 0 ]
-}
-
