+++ test/doc.go
@@ -1,2 +1,2 @@
 // Package test contains test definitions, code, and data.
-package test
+package main
