--- test/features/cli.feature
@@ -1,8 +1,0 @@
-Feature: Simple CLI commands
-
-    @Acceptance
-    Scenario: Prints help correctly
-        When the app runs with parameters "-h"
-        Then the app exits without error
-        And the app output contains "Usage:"
-
