+++ test/features/tests.feature
@@ -1,0 +1,2 @@
+Feature: Error cases
+
