+++ test/steps_test.go
@@ -1,23 +1,26 @@
-package test_test
+//go:build acceptance
+
+package main_test
 
 import (
 	"fmt"
+	"os"
 	"os/exec"
 	"strings"
 
-	"github.com/cucumber/godog"
 	"github.com/stretchr/testify/assert"
 )
 
-//nolint: unused
+// nolint: unused
 type testContext struct {
 	err      error
 	cmdInput struct {
 		parameters string
 	}
 	cmdResult struct {
-		Output string
-		Err    error
+		Output       string
+		Err          error
+		ProcessState *os.ProcessState
 	}
 }
 
@@ -30,17 +33,13 @@ func (c *testContext) Errorf(format string, args ...interface{}) {
 func (c *testContext) theAppRunsWithParameters(args string) error {
 	c.cmdInput.parameters = args
 	cmdArgs := strings.Split(args, " ")
-	cmd := exec.Command("../goclitem", cmdArgs...)
+	cmd := exec.Command("./testapp", cmdArgs...)
 	output, err := cmd.CombinedOutput()
 	c.cmdResult.Output = string(output)
 	c.cmdResult.Err = err
+	c.cmdResult.ProcessState = cmd.ProcessState
 
 	return nil
-}
-
-func (c *testContext) theAppExitsWithoutError() error {
-	assert.NoError(c, c.cmdResult.Err)
-	return c.err
 }
 
 func (c *testContext) theAppExitsWithAnError() error {
@@ -58,30 +57,4 @@ func (c *testContext) theAppOutputDoesNotContain(unexpected string) error {
 	unexpected = strings.ReplaceAll(unexpected, "\\\"", "\"")
 	assert.NotContains(c, c.cmdResult.Output, unexpected)
 	return c.err
-}
-
-//nolint: unused
-func InitializeTestSuite(ctx *godog.TestSuiteContext) {
-	ctx.BeforeSuite(func() {})
-}
-
-//nolint: unused
-func InitializeScenario(ctx *godog.ScenarioContext) {
-	tc := testContext{}
-	ctx.BeforeScenario(func(*godog.Scenario) {})
-	ctx.AfterScenario(func(s *godog.Scenario, err error) {
-		if err != nil {
-			fmt.Printf(
-				"Command line output for \"%s\"\nUsing parameters: %s\n%s",
-				s.GetName(),
-				tc.cmdInput.parameters,
-				tc.cmdResult.Output,
-			)
-		}
-	})
-	ctx.Step(`^the app runs with parameters "(.*)"$`, tc.theAppRunsWithParameters)
-	ctx.Step(`^the app exits without error$`, tc.theAppExitsWithoutError)
-	ctx.Step(`^the app exits with an error$`, tc.theAppExitsWithAnError)
-	ctx.Step(`^the app output contains "(.*)"$`, tc.theAppOutputContains)
-	ctx.Step(`^the app output does not contain "(.*)"$`, tc.theAppOutputDoesNotContain)
 }
